+++ billingo/src/Enums/OnlinePaymentEnum.php
@@ -7,5 +7,7 @@ enum OnlinePaymentEnum: string
     case EMPTY = '';
     case BARION = 'Barion';
     case SIMPLEPAY = 'SimplePay';
+    case online_bankcard = 'online_payment';
+    case online_payment = 'Stripe';
     case NO = 'no';
 }